Restrict the maximum number of columns/rows
Running Gargoyle with either a high resolution or a small font reveals
buffer overflows due to buffers that are defined as char[256].

This patch works around this problem by limiting the number of text
columns and rows to 255. The margin is increased when the window is
larger.

The Gtk version previously limited the maximum window size to 255
columns and 250 rows, which interfered with the fullscreen mode.
This limit has been removed.